Moved suspension tracking to `IORuntime`
